// generated/src/app/api-models/al-letter-template-dtl.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AlLetterTemplateDtl {
  seqLetterGroupId: number;
  seqLetterTempId: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}